Move API_OPTIONS to lib. Disable caching.
